Rename byte_reader file in byte_read
