+++ src/utils/mod.rs
@@ -1,1 +1,1 @@
-pub mod byte_reader;
+pub mod byte_read;
